Work on tests.
- Renamed test
- Added test for percentages
