+++ reports/gp2gp_transfer_status_report/gp2gp_transfer_status_report_snapshot_percentages.splunk
@@ -1,0 +1,16 @@
+index="$index$"
+
+{% from 'gp2gp_transfer_status_report/gp2gp_transfer_status_report_snapshot_base.splunk' import
+gp2gp_transfer_status_report_snapshot_base %}
+
+
+{{ gp2gp_transfer_status_report_snapshot_base() }}
+
+
+| table 
+total_eligible_for_electronic_transfer,  
+percentage_successfully_integrated,
+percentage_rejected, 
+percentage_awaiting_integration, 
+percentage_in_progress,
+percentage_technical_failure
